requirements.txt
Django==3.2
djangorestframework==3.12.4
coolname==1.1.0
Faker==8.1.1
